Dead-code removal, clippy cleanup, and hygiene fixes in mangler_core
- Delete dead dynamic_image_serde module (superseded by float_image_serde)
- Delete unused ConnectionSettings/UiType, PathType, UiButton, and the
  commented-out text_from_clipboard stub
- Remove never-set Node.is_busy and write-only Graph.is_dirty
- Fix mislabeled send-error logs in node.rs via try_send_node_changed helper
- Make Value::fingerprint allocation-free (hash inner enum discriminants
  instead of format!) and add enum-distinctness test
- Fix clippy warnings across core and operations (approx_constant,
  module_inception, print_literal, ptr_arg, clone_on_copy, collapsible_match)
- Remove unused tiny.mp4 CLI test fixture; add OFL.txt font licenses

+++ app/crates/mangler_core/src/operations/images/adjustments/gradient_dynamic.rs
@@ -91,7 +91,7 @@ impl OpImageAdjustmentGradientDynamic {
         let fx_scale = if field_w > 0 { field_w as f32 / width.max(1) as f32 } else { 0.0 };
         let fy_scale = if field_h > 0 { field_h as f32 / height.max(1) as f32 } else { 0.0 };
 
-        let angle_rad = (angle as f32).to_radians();
+        let angle_rad = angle.to_radians();
         let cos_a = angle_rad.cos();
         let sin_a = angle_rad.sin();
