+++ app/crates/mangler_core/src/operations/images/filter/erode.rs
@@ -65,7 +65,7 @@ impl OpImageAdjustmentErode {
         let Value::Image { data, change_id: _ } = image_converted.unwrap() else { unreachable!() };
         let Value::Integer(radius) = radius_converted.unwrap() else { unreachable!() };
 
-        let radius = radius.max(1) as i32;
+        let radius = radius.max(1);
 
         // Separable min-filter: horizontal pass then vertical pass. A square
         // min kernel factors into 1D min ops, reducing cost from O(r²) to O(r).
