+++ app/crates/mangler_core/src/operations/images/filter/morphological_gradient.rs
@@ -61,7 +61,7 @@ impl OpImageAdjustmentMorphGradient {
 
         let Value::Image { data, change_id: _ } = image_converted.unwrap() else { unreachable!() };
         let Value::Integer(radius) = radius_converted.unwrap() else { unreachable!() };
-        let radius = radius.max(1) as i32;
+        let radius = radius.max(1);
 
         let (w, h) = data.dimensions();
         let dilated = separable_morphology(&data, radius, |a, b| a.max(b));
